Draft: Split Sketch from Draft.py
